Extract RamPanel page object for CPU and VM RAM access
Collapses near-identical "select dec, fill Addr, press Enter, read sibling
cell" code in CpuPage.readRamCell/setRamCell and VmPage.readRam into a
single RamPanel page object. CpuPage and VmPage keep thin pass-through
methods so existing spec call sites are unchanged. The chip-tool RAM read
path uses bus inputs (different DOM) and is intentionally not collapsed.
