+++ e2e/src/pages/CpuPage.ts
@@ -1,11 +1,14 @@
 import { expect, type Page } from "@playwright/test";
 import { FilePicker } from "./FilePicker";
+import { RamPanel } from "./RamPanel";
 
 export class CpuPage {
   readonly filePicker: FilePicker;
+  readonly ramPanel: RamPanel;
 
   constructor(private _page: Page) {
     this.filePicker = new FilePicker(_page);
+    this.ramPanel = new RamPanel(_page.locator("article.panel.memory.RAM"));
   }
 
   get page(): Page {
@@ -30,43 +33,16 @@ export class CpuPage {
     await this.filePicker.selectPath(pathSegments);
   }
 
-  private get ramPanel() {
-    return this._page.locator("article.panel.memory.RAM");
-  }
-
   private get ioPanel() {
     return this._page.locator("article.panel.IO");
   }
 
-  private async ramRow(address: number) {
-    await this.ramPanel.locator("select").selectOption("dec");
-    const addrInput = this.ramPanel.getByPlaceholder("Addr");
-    await addrInput.fill(String(address));
-    await addrInput.press("Enter");
-    const indexCell = this.ramPanel
-      .locator("code:first-child")
-      .filter({ hasText: new RegExp(`^\\s*${address}\\s*$`) })
-      .first();
-    await expect(indexCell).toBeVisible();
-    return indexCell;
-  }
-
   async readRamCell(address: number): Promise<number> {
-    const indexCell = await this.ramRow(address);
-    const value = await indexCell.evaluate(
-      (el) => el.nextElementSibling?.textContent?.trim() ?? "",
-    );
-    return Number(value);
+    return this.ramPanel.readAt(address);
   }
 
   async setRamCell(address: number, value: number): Promise<void> {
-    const indexCell = await this.ramRow(address);
-    const valueCell = indexCell.locator("xpath=following-sibling::code[1]");
-    await valueCell.click();
-    const input = valueCell.locator("input");
-    await input.fill(String(value));
-    await input.press("Enter");
-    await expect.poll(async () => await this.readRamCell(address)).toBe(value);
+    await this.ramPanel.writeAt(address, value);
   }
 
   async step(): Promise<void> {
