+++ e2e/src/pages/VmPage.ts
@@ -1,11 +1,14 @@
 import { expect, type Locator, type Page } from "@playwright/test";
 import { FilePicker } from "./FilePicker";
+import { RamPanel } from "./RamPanel";
 
 export class VmPage {
   filePicker: FilePicker;
+  ramPanel: RamPanel;
 
   constructor(private _page: Page) {
     this.filePicker = new FilePicker(_page);
+    this.ramPanel = new RamPanel(_page.locator("article.panel.memory.RAM"));
   }
 
   get page(): Page {
@@ -50,20 +53,7 @@ export class VmPage {
   }
 
   async readRam(address: number): Promise<number> {
-    const ramPanel = this._page.locator("article.panel.memory.RAM");
-    await ramPanel.locator("select").selectOption("dec");
-    const addrInput = ramPanel.getByPlaceholder("Addr");
-    await addrInput.fill(String(address));
-    await addrInput.press("Enter");
-    const indexCell = ramPanel
-      .locator("code:first-child")
-      .filter({ hasText: new RegExp(`^\\s*${address}\\s*$`) })
-      .first();
-    await expect(indexCell).toBeVisible();
-    const value = await indexCell.evaluate(
-      (el) => el.nextElementSibling?.textContent?.trim() ?? "",
-    );
-    return Number(value);
+    return this.ramPanel.readAt(address);
   }
 
   async fillEditor(content: string): Promise<void> {
